event edit

+++ app/Http/Controllers/instituteController.php
@@ -81,13 +81,63 @@ class instituteController extends Controller
         $user = Auth::user();
         $date = date('Y-m-d');
         $event = event::find($eventId);
-        if($event->user_id == $user->id){
-            if($event->startDate > $date){
-                $event->delete();
-            }
-        }   
+        if($event){
+            if($event->user_id == $user->id){
+                if($event->startDate > $date){
+                    $event->delete();
+                }
+            }   
+        }
+        return redirect()->route('myEvents');
+    }
+
+    public function eventVeiwEdit($eventId){
+        $user = Auth::user();
+        $followers = friend::where('requested_id', $user->id);
+        $following = friend::where('requester_id', $user->id);
+        $date = $this->date;
+        $Aevents = event::where('user_id', $user->id)->where('startDate','<',$date);
+        $Uevents = event::where('user_id', $user->id)->where('startDate','>',$date);
+        $event = event::find($eventId);
+        if($event){
+            if($event->user_id == $user->id){
+                if($event->startDate > $date){
+                    return view('institute/eventEdit',compact('event','user','Aevents','Uevents','followers','following'));
+                }
+            }   
+        }
+        return redirect()->route('home');
+    }
+
+    public function eventEdit(Request $request){
+        $user = Auth::user();
+        $this->validate($request, [
+            'eventId' => 'required',
+            'title' => 'required|string|max:100',
+            'description' => 'required|string',
+            'startDate' => 'required|date|after:tomorrow',
+            'endDate' => 'required|date|after:start_date',
+        ]);
+
+        $eventId = $request['eventId'];
+        $event = event::find($eventId);
+        if($event){
+            if($event->user_id == $user->id){
+                if($event->startDate > $date){
+                    $event->title = $request['title'];
+                    $event->user_id = $user->id;
+                    $event->description = $request['description'];
+                    $event->startDate = $request['startDate'];
+                    $event->endDate = $request['endDate'];
+                    $event->save();
+                    return redirect()->route('event',compact('event'));
+                }
+            }   
+        }
         return redirect()->back();
     }
+
+
      public function findVolunteers()
     {
             $user = Auth::user();
